Upgrade to latest github-pages and enable Solidity syntax highlighting (#2679)
After https://github.com/github/pages-gem/pull/683 / https://github.com/github/pages-gem/pull/690 was merged this is possible now.

This allows EIPs to use solidity syntax highlighting. Currently there are 330 instances of it:

```
$ grep "\`solidity" * | wc -l
     330
```

And some EIPs used `typescript` or `javascript` as an alternative, because that was highlighted already:
```
$ grep "\`typescript" * | wc -l
      16
$ grep "\`javascript" * | wc -l
      25
```

Thanks @veox again in adding highlighting to rouge.