Fixes "spelling error" in diff (#5105)
I was hoping that this file would fail once, not every time someone submits a PR.  Unfortunately I was wrong.  I have no idea if this will invalidate the patch, but I'm optimistic that the diff is lenient on line lengths and having a few more characters than the diff's default line length won't break anything.  😬