Add EIP-5773: Multi-Resource Token standard (#5773)
* Propose Multi-Resource Token standard

RMRK team has developed a next step in NFTs where one NFT can be tied to
multiple resources.

* Address issues reported by EIP repository's CI

* Fix styling discrepancies in section titles

* Address CI's issues in the preabmle

* Add discussion URL

* Fix discussion URL formatting

* Add the EIP number

* Fix getResourceMeta specification

* Apply changes based on PR comments

* Replace "primitive" with a clearer explanation

Co-authored-by: Pandapip1 <45835846+Pandapip1@users.noreply.github.com>

* Apply changes based on ECH call #5

Applied the changes based on feedback received on the EIP Editing Office
Hour Meeting #5.

This commit includes changes to the proposal as well as adds an exaple
implementation to assets/ directory.

* Address the stylistic issues reported by the CI

* Fix a reference to EIP-712

* Minor typo fixes

* Updates `getResourceMetadata` description

Co-authored-by: Sam Wilson <57262657+SamWilsn@users.noreply.github.com>

* Rename MultiResource -> MultiAsset and relicense

The proposal was renamed to Context-Dependent Multi-Asset Tokens to
better illustrate its function.

Another example was added to represent the possible IoT usecase and the
explanation on the naming decision was added to the rationale.

The examples were relicensed to CC0, to conform to the requirements of
EIP repository.

* Fix license identifier & Specification contract title

* Fix indexes of Rationale questions

* Minor polishing of the rationale

* Update references to assets to use an istead of a in Rationale

* Final polishes

Co-authored-by: Pandapip1 <45835846+Pandapip1@users.noreply.github.com>
Co-authored-by: Steven Pineda <steven2308@outlook.com>
Co-authored-by: Sam Wilson <57262657+SamWilsn@users.noreply.github.com>