Add EIP-6059: Parent-Governed Nestable Non-Fungible Tokens standard (#6059)
* Propose Parent-Governed Nestable Non-Fungible Tokens standard

RMRK team has developed a next step in NFT evolution where one NFT can
own and manage other NFTs.

* Assign EIP number

EIP number 6059 was assigned as this is the number of the PR to add this
proposal.

This was doen before being assigned the EIP number by one of the editors
based on the common practice of assigning EIP numbers based on the PR
number at which the proposal was added. In case the number of the
proposal should be different, the number can easiliy be changed.

* Add discussion URI and minor text-fixes

The discussion URI was added and some minor text fixes addressed.

* Update examples references

* Update EIPS/eip-6059.md

Uses "account" instead of "EOA"

Co-authored-by: Sam Wilson <57262657+SamWilsn@users.noreply.github.com>

* Apply grammar suggestions from code review

Co-authored-by: Sam Wilson <57262657+SamWilsn@users.noreply.github.com>

Co-authored-by: Steven Pineda <steven2308@outlook.com>
Co-authored-by: Sam Wilson <57262657+SamWilsn@users.noreply.github.com>